requirements.txt
anyio==3.6.1
APScheduler==3.9.1
beautifulsoup4==4.11.1
black==22.6.0
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
feedparser==6.0.10
h11==0.12.0
httpcore==0.15.0
httpx==0.23.0
idna==3.3
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.2
pymongo==4.1.1
python-dateutil==2.8.2
python-dotenv==0.20.0
python-telegram-bot==20.0a2
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
requests==2.28.1
rfc3986==1.5.0
sgmllib3k==1.0.0
six==1.16.0
sniffio==1.2.0
soupsieve==2.3.2.post1
tomli==2.0.1
tornado==6.1
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.9
